Revert "Merge pull request #50 from taranpalsingh/develop"
This reverts commit 34b28d43fc957e314f432f5d81eab2c981a881b8, reversing
changes made to cd1f1247f3d17e06b09e511c173049ec787c707f.
