+++ ms-dashboard/package.json
@@ -25,7 +25,6 @@
     "@angular/router": "8.0.2",
     "@ng-bootstrap/ng-bootstrap": "5.0.0",
     "@ngui/map": "0.30.3",
-    "angular-file-uploader": "^5.0.2",
     "arrive": "2.4.1",
     "bootstrap": "4.3.1",
     "bootstrap-notify": "3.1.3",
@@ -38,7 +37,6 @@
     "hammerjs": "^2.0.8",
     "jquery": "3.4.1",
     "moment": "2.24.0",
-    "ng-pick-datetime": "^7.0.0",
     "ngx-toastr": "10.0.4",
     "rxjs": "6.5.2",
     "sweetalert2": "^9.4.2",
