+++ ms-dashboard/src/app/app.module.ts
@@ -16,7 +16,7 @@ import { FixedPluginComponent } from './shared/Components/fixedplugin/fixedplugi
 import { LoginComponent } from './user/login/login.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule, MatInputModule, MatButtonModule, MatIconModule, MatCheckboxModule, MatDialogModule, MatDatepickerModule, MatNativeDateModule, MatSelectModule, MatTabsModule } from '@angular/material';
-// import { NewpostComponent } from './pages/newpost/newpost.component';
+import { NewpostComponent } from './newpost/newpost.component';
 import { NewpostDialogComponent } from './newpost-dialog/newpost-dialog.component';
 import { TwitterAuthComponent } from './auth/twitter-auth/twitter-auth.component';
 import { HttpClientModule } from '@angular/common/http';
@@ -29,7 +29,7 @@ import { ChangePasswordComponent } from './user/change-password/change-password.
     AppComponent,
     AdminLayoutComponent,
     LoginComponent,
-    // NewpostComponent,
+    NewpostComponent,
     NewpostDialogComponent,
     TwitterAuthComponent,
     EmailVerificationComponent,
