+++ ms-dashboard/src/app/layouts/admin-layout/admin-layout.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormsModule } from '@angular/forms';
 
 import { AdminLayoutRoutes } from './admin-layout.routing';
 
@@ -10,11 +10,7 @@ import { UserComponent }            from '../../pages/user/user.component';
 import { WordCloudComponent } from '../../charts/word-cloud/word-cloud.component';
 
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { MatFormFieldModule, MatInputModule, MatButtonModule, MatIconModule, 
-  MatCheckboxModule, MatDialogModule, MatDatepickerModule, MatNativeDateModule, MatSelectModule, MatTabsModule, MatCardModule, MatSliderModule, MatSlideToggleModule } from '@angular/material';
-import { NewpostComponent } from 'app/pages/newpost/newpost.component';
-import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'ng-pick-datetime';
-import { AngularFileUploaderModule } from "angular-file-uploader";
+import { MatFormFieldModule, MatInputModule, MatButtonModule, MatIconModule, MatCheckboxModule, MatDialogModule, MatDatepickerModule, MatNativeDateModule, MatSelectModule, MatTabsModule } from '@angular/material';
 import { SocialAccountLoginComponent } from 'app/pages/social-account-login/social-account-login.component';
 
 @NgModule({
@@ -32,19 +28,12 @@ import { SocialAccountLoginComponent } from 'app/pages/social-account-login/soci
     MatDatepickerModule,
     MatNativeDateModule,
     MatSelectModule,
-    MatTabsModule,
-    ReactiveFormsModule,
-    OwlDateTimeModule,
-    OwlNativeDateTimeModule,
-    MatCardModule,
-    AngularFileUploaderModule,
-    MatSlideToggleModule
+    MatTabsModule
   ],
   declarations: [
     DashboardComponent,
     UserComponent,
     WordCloudComponent,
-    NewpostComponent,
     SocialAccountLoginComponent
   ],
   entryComponents:[
