+++ ms-dashboard/src/app/layouts/admin-layout/admin-layout.routing.ts
@@ -3,7 +3,6 @@ import { Routes } from '@angular/router';
 import { DashboardComponent } from '../../pages/dashboard/dashboard.component';
 import { UserComponent } from '../../pages/user/user.component';
 import { AuthGuardService } from 'app/auth-guard.service';
-import { NewpostComponent } from 'app/pages/newpost/newpost.component';
 
 export const AdminLayoutRoutes: Routes = [
     { 
@@ -21,8 +20,4 @@ export const AdminLayoutRoutes: Routes = [
         component: UserComponent,     
         // canActivate: [AuthGuardService]
     },
-    {
-        path: 'new',
-        component: NewpostComponent
-    }
 ];
