+++ ms-dashboard/src/app/newpost-dialog/newpost-dialog.component.html
@@ -7,9 +7,12 @@
         <div class="form-group">
             <mat-form-field class="example-full-width textarea">
                 <textarea matInput name="Text" cols="40" rows="5" 
-                    placeholder="Text" [(ngModel)]="data.text" >
+                 placeholder="Text" [(ngModel)]="data.text" >
                 </textarea>
-            </mat-form-field>
+               </mat-form-field>
+            <!-- <textarea name="Text" cols="40" rows="15" 
+                class="form-control" placeholder="Text">
+            </textarea> -->
         </div>
         <div class="form-group">
             <mat-checkbox color="primary" name="postNow"
