+++ ms-dashboard/src/app/newpost/newpost.component.html
@@ -1,0 +1,15 @@
+<h2>Publish a new Post</h2>
+<form #f="ngForm" (ngSubmit)="Submit(f)">
+    <div class="form-group">
+        <mat-form-field>
+          <input required matInput placeholder="Frst Name" [value]="firstname" 
+                 name="firstname" ngModel>
+        </mat-form-field>
+    </div>
+    <div *ngIf="isSignup" class="form-group">
+      <mat-form-field>
+        <input required matInput placeholder="Last Name" [value]="lastname" 
+               name="lastname" ngModel>
+      </mat-form-field>
+    </div>
+</form>
