+++ ms-dashboard/src/app/newpost/newpost.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-newpost',
+  templateUrl: './newpost.component.html',
+  styleUrls: ['./newpost.component.scss']
+})
+export class NewpostComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
