--- ms-dashboard/src/app/pages/newpost/newpost.component.html
@@ -1,113 +1,0 @@
-<div class="container text-center">
-  <div class="row">
-    <div class="col-md-6">
-      <h2 class="text-center my-4">New Post!!</h2>
-      <div class="icons my-5">
-        <i class="fa fa-twitter-square media" 
-          [class.media-selected]="twitterSelected" 
-          (click)="twitterSelected=!twitterSelected">
-        </i>
-        <i class="fa fa-facebook-square media" 
-          [class.media-selected]="facebookSelected"
-          (click)="facebookSelected=!facebookSelected">
-        </i>
-        <i class="fa fa-instagram media" 
-          [class.media-selected]="instaSelected"
-          (click)="instaSelected=!instaSelected">
-        </i>
-      </div>
-      <form [formGroup]="newPostFormGroup" (ngSubmit)="submitForm()">
-        <div class="form-group">
-          <mat-form-field class="text">
-            <textarea matInput formControlName="text" rows="3" 
-                placeholder="Text" maxlength="140">
-            </textarea>
-            <mat-error>
-              <span *ngIf="
-                  !newPostFormGroup.get('text').valid &&
-                  newPostFormGroup.get('text').touched
-                ">
-                Please enter valid Text.
-              </span>
-            </mat-error>
-          </mat-form-field>
-        </div>
-    <!-- ///////////////////////// -->
-    
-        <!-- <angular-file-uploader class="text-center"
-          [config]="afuConfig">
-        </angular-file-uploader> -->
-    
-        <!-- <angular-file-uploader
-          [config]="afuConfig"
-          [resetUpload]=resetVar
-          (ApiResponse)="DocUpload($event)"
-          >
-        </angular-file-uploader> -->
-    <!-- ///////////////////////// -->
-    
-        <!-- <div class="form-group">
-          <mat-checkbox color="primary" formControlName="scheduleIt">
-              <span>Schedule</span>
-          </mat-checkbox>
-    
-          formControlName="scheduleIt" 
-          (click)="newPostFormGroup.value.scheduleIt=!newPostFormGroup.value.scheduleIt" 
-    
-        </div> -->
-        <!-- <button color="primary" mat-stroked-button
-            (click)="dateTime.click()" 
-            type="button">
-            Schedule It 
-        </button> -->
-    
-        <div class="dragNdrop mx-auto mb-5" (click)="imageuploaded=!imageuploaded;"></div>
-
-        <div class="schedule">
-          <div class="form-group">
-            <mat-slide-toggle color="primary" formControlName="scheduleIt">
-              Schedule It!
-            </mat-slide-toggle>
-          </div>    
-          <div *ngIf="newPostFormGroup.value.scheduleIt" class="form-group">
-            <mat-form-field>
-              <input #dateTime name="dateTime" [disabled]="newPostFormGroup.value.scheduleIt" matInput formControlName="dateTime" [owlDateTime]="dt1" 
-                [owlDateTimeTrigger]="dt1" placeholder="Date & Time">
-              <owl-date-time #dt1></owl-date-time>
-            </mat-form-field>
-          </div>  
-        </div>
-        
-        <button mat-stroked-button type="submit"
-          color="primary" class="my-3">
-          Create
-        </button>
-      </form>        
-    </div>
-    <div class="col-md-6">
-      <h2 class="text-center my-4">Preview</h2>
-      <mat-card class="twitter-card mx-auto my-5">
-        <mat-card-header>
-          <div mat-card-avatar class="example-header-image">
-            <img src="https://material.angular.io/assets/img/examples/shiba2.jpg" 
-              alt="Photo of a Shiba Inu" class="profile-img mt-2">
-          </div>
-          <mat-card-title>Username</mat-card-title>
-          <mat-card-subtitle class="text-left">@screen_name</mat-card-subtitle>
-        </mat-card-header>
-        <mat-card-content>
-          <p class="text-left">{{newPostFormGroup.value.text}}</p>
-        </mat-card-content>
-        <img mat-card-image  *ngIf="imageuploaded" src="https://material.angular.io/assets/img/examples/shiba2.jpg" 
-            alt="Photo of a Shiba Inu" class="image_upload">
-        <mat-card-actions>
-          <div class="twitterActions mb-3">
-            <i class="fa fa-comment-o"></i>
-            <i class="fa fa-retweet"></i>  
-            <i class="fa fa-heart-o"></i>
-          </div>
-        </mat-card-actions>
-      </mat-card>
-    </div>
-  </div>
-</div>
