+++ ms-dashboard/src/app/shared/Components/navbar/navbar.component.html
@@ -35,9 +35,10 @@
             </p>
           </a>
         </li>
+
         <li class="nav-item">
           <a class="nav-link btn-magnify" href="javascript:void(0)" 
-              matTooltip="New Post" routerLink="/new">
+              matTooltip="New Post" (click)="openDialog()">
             <i class="nc-icon nc-simple-add"></i>
             <p>
               <span class="d-lg-none d-md-block">New Post</span>
