+++ ms-dashboard/src/index.html
@@ -1,36 +1,36 @@
 <!doctype html>
 <html>
-    <head>
-        <meta charset="utf-8" />
-        <link rel="apple-touch-icon" sizes="76x76" href="./assets/img/logo.png">
-        <link rel="icon" type="image/png" sizes="96x96" href="./assets/img/logo.png">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-        <meta content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0' name='viewport' />
-        <meta name="viewport" content="width=device-width" />
+  <head>
+      <meta charset="utf-8" />
+      <link rel="apple-touch-icon" sizes="76x76" href="./assets/img/logo.png">
+      <link rel="icon" type="image/png" sizes="96x96" href="./assets/img/logo.png">
+      <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
+      <meta content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0' name='viewport' />
+      <meta name="viewport" content="width=device-width" />
 
-        <title>Socialize</title>
+      <title>Socialize</title>
 
-        <base href="/">
+      <base href="/">
 
-        <!--     Fonts and icons     -->
-        <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700,200" rel="stylesheet" />
-        <link href="https://maxcdn.bootstrapcdn.com/font-awesome/latest/css/font-awesome.min.css" rel="stylesheet">
+      <!--  CSS for Demo Purpose, don't include it in your project     -->
+      <link href="./assets/css/demo.css" rel="stylesheet" />
 
-        <!--     file-upload     -->
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+      <!--     Fonts and icons     -->
+      <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700,200" rel="stylesheet" />
+      <link href="https://maxcdn.bootstrapcdn.com/font-awesome/latest/css/font-awesome.min.css" rel="stylesheet">
 
-        <!-- Font Awesome -->
-        <!-- <link rel="stylesheet" href="path/to/font-awesome/css/font-awesome.min.css"> -->
+      <!-- Font Awesome -->
+      <!-- <link rel="stylesheet" href="path/to/font-awesome/css/font-awesome.min.css"> -->
 
-        <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    </head>
-    <body>
-        <app-root>
-            <div class="loader-container">
-                <div class="loader"></div>
-            </div>
-        </app-root>
-    </body>
+      <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet">
+      <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+</head>
+  <body>
+      <app-root>
+          <div class="loader-container">
+              <div class="loader"></div>
+          </div>
+      </app-root>
+  </body>
 
 </html>
